Modificado Producto Controller
Modificada la clase ProductoController

+++ ljorda-micro-comparador-precios/src/main/java/es/luciajorda/comparadorprecios/rest/ProductoController.java
@@ -1,11 +1,10 @@
 package es.luciajorda.comparadorprecios.rest;
 
+import java.net.URI;
 import java.util.List;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import es.luciajorda.comparadorprecios.domain.Producto;
 import es.luciajorda.comparadorprecios.service.ProductoService;
@@ -25,32 +25,65 @@ public class ProductoController {
 	private ProductoService productoService;
 	
 	@GetMapping("/producto")
-	public List<Producto> getAll() {
-		return this.productoService.getAll();
+	public ResponseEntity <List<Producto>> getAll() {
+		List<Producto> listaProductos = this.productoService.getAll();
+		if(listaProductos == null) {
+			return ResponseEntity.notFound().build();
+		}
+		return ResponseEntity.ok().body(listaProductos);
 	}
 	
 	@GetMapping("/producto/{id}")
-	public Producto getProductById(@PathVariable("id") Long id) {
-		return this.productoService.getById(id);
+	public ResponseEntity<Producto> getById(@PathVariable("id") Long id) {
+		Producto producto = this.productoService.getById(id);
+		if(producto == null) {
+			return ResponseEntity.notFound().build();
+		}
+		return ResponseEntity.ok().body(producto);
 	}
 	
 	@PostMapping("/producto")
-	public void createProducto(@RequestBody Producto producto, HttpServletRequest request, HttpServletResponse response) {
-		response.setHeader("Location", request.getRequestURL().
-				append("/").
-				append(this.productoService.create(producto).getId()).toString());
+	public ResponseEntity<Producto> create(@RequestBody Producto producto) {
+		Producto createProducto = this.productoService.create(producto);
+		
+		URI location = ServletUriComponentsBuilder
+							.fromCurrentContextPath()
+							.path("/id")
+							.buildAndExpand(createProducto.getId())
+							.toUri();
+		
+		return ResponseEntity.created(location).body(createProducto);
 	}
 	
 	@PutMapping("/producto/{id}")
-	public void updateProducto(@PathVariable("id") Long id,@RequestBody Producto producto) {
-		Producto update = this.productoService.getById(id);
-		update.setName(producto.getName());
-		this.productoService.create(update);
+	public ResponseEntity<Producto> update(@PathVariable("id") Long id,@RequestBody Producto producto) {
+		Producto updateProducto = this.productoService.getById(id);
+		
+		if(updateProducto == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		updateProducto.setName(producto.getName());
+		this.productoService.create(updateProducto);
+		
+		URI location = ServletUriComponentsBuilder
+							.fromCurrentRequest()
+							.build()
+							.toUri();
+		
+		return ResponseEntity.created(location).body(producto);
 	}
 	
 	@DeleteMapping("/producto/{id}")
-	public void deleteProducto(@PathVariable("id") Long id) {
+	public ResponseEntity<Producto> delete(@PathVariable("id") Long id) {
+		if(this.productoService.getById(id) == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
 		this.productoService.delete(id);
+		
+		return ResponseEntity.ok().build();
+			
 	}
 
 }
